src/agent/pyproject.toml
[project]
name = "company-intel-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim[ag-ui,openai]",
    "fastapi",
    "uvicorn[standard]",
    "logfire[fastapi,httpx]",
    "crawl4ai",
    "langdetect",
    "duckduckgo-search",
    "qdrant-client>=1.12",
    "fastembed>=0.4",
    "tiktoken>=0.8",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["crawl4ai.*", "langdetect.*", "fastembed.*", "qdrant_client.*", "tiktoken.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
